requirements.txt
absl-py==1.4.0
aioice==0.7.6
aiortc==1.4.0
altair==4.2.2
astor==0.8.1
astunparse==1.6.3
attrs==22.2.0
av==10.0.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.2
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==7.1.2
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
facenet-pytorch==2.5.2
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
gdown==4.6.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
keras-facenet==0.3.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mtcnn==0.1.1
netifaces==0.11.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
protobuf==3.19.6
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
streamlit==1.3.1
streamlit-webrtc==0.11.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
unzip==1.0.0
urllib3==1.26.14
validators==0.20.0
watchdog==2.3.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
